[WIP] Refactor to use CommandLineUtils and change Shortcut Form
Changed to use natemcmaster's CommandLineUtils library to provide
flexbility for command structure, and enable the use of custom
validators which allow us to make sure that the command line options
provided are accurate. This removes the need for all the individual
exceptions being thrown across the main application.
    Also started the modification of the CreateShortcut Form to make
it easier to use and more self explanatory.
    And finally started the integration of Uplay into the application so
that users can use Uplay as well as Steam to select their games.