[WIP] Modified shortcut generation and Uplay skeleton
The programme compiles with the commandline changes
and the changes to the shortcut form, including aligning
the shortcut generation with the new command line.
   The Uplay skeleton has been disabled for now as I'm
concentrating on getting the commandline args working
and fixing the annoying errors in the way the MainForm
and the ShortcutForm work. Probably have a couple more
days work to fix that.