Created new Main Form to separate main tasks
It seemed to me that it would make more sense to
separate tht two main tasks completely and have them
independent from each other in the program. So I
created a new MainForm that opens up either a
'Setup Display Profiles' or 'Setup Game Shortcuts'
window depending on what the user clicks.
    The idea is that Display Profile setup only happens
when the app is first installed, and after that it's just
the game shortcut that gets updated.
    Also created a first attempt at a new HeliosPlus Icon.