[WIP] Started dev of the ImagelistView ProfileAdaptor
Started to work on getting the profile images directly
into the imagelistview rather than doing it via files.
Wasn't a fan of doing icons by file but it was the fastest
way to getting things going. Now going to try and sort
it out so I feel a bit bettter about the code.