[WIP] Testing logic for detecting displays
The existing logic for detecting displays connected to
an NVidia device doesn't allow matching for Nvidia
surround and normal surroundless profiles. This means
that you cannot determine which profiles are the valid
for your display setup. This is testing the NVAPIWRapper
scanning and the WindowsDisplayAPI libraries to see what
information will work for tracking the screen in the profiles.