Fixed WindowsDisplayAPI Device Identifcation
Fixed the WindowsDisplayAPI Device Identifcation
logic so that it reliably detects the displays in a way
that works when some displays are disabled as well
as when they are enabled. Also divided up the
DeviceIdentification into two separate functions so
that one will list all attached devices, and the other
will only report the devices currently in use.