First working surround to non-surround conversion
Was missing a SetTopologies  command from upstream
and the latest GeForce devices don't seem to be able to
read EDID which means the Topologies dont have a
DisplayName peroperty, which in turn causes issues when
trying to change to that topology! So now detect empty
DisplayName and fill it with a default.