Correctly close programs that we start
The StartPrograms processes now all correctly are added
to a list of processes to stop only if the user requested they
are stopped.

Additionally, a 300ms delay was added after SteamGame is
launched to give it time to actually run the darn game! It was
erroring due to testing for the launch too quickly. Steam
takes a wee bit of time to start the game, and we need to
account for that.