[WIP] populated dropdown with audio devices
Part of the work through getting audio devices
into DisplayMagician, as it now scans the devices
and automatically selects the one in use currently.
You can also press the 'rescan' button to have it
update the list. This is the UI part done.

Next step is the saving the chosen audio device
to the saved Shortcut!