Fixed starting all start programs
This fixes an edge case where a start program
is already running and DM would skip starting
any additional start programs. Fixes #18.

Also fixed an issue with it trying to stop start
programs processes that were already stopped
by the user or by some other process. It now
logs the fact this wasn't done as it was already
closed.