[WIP] Working nearly DisplayIds
Have now go the required data, but for some
reason AMD is returning two valid adapters
when there is actually just one in the system!

Could be a problem with the device installation,
but I will need figuring out how to either remove
duplicates, or figure out the information to use as
a filter for ignoring one of the adapters.