Fixed exceprion with no game shortcut
Fixes #25. When you attempted to create a shortcut
that didnt run a game or executable, the program would
save the shortcut, but would crash when trying to show
it in the Shortcut library. This was because the _shortcutBitmap
wsn't being populated when no game or executable was
chosen. This has been fixed now, as it will use the profile bitmap
as the shortcut bitmap in these instances.