Removed LogReporter and updated ShellExtension
Removed LogReporter as DisplayMagician now contains most of the required logging that I need to actually fix things.

Updated the ShellExtension as it was looking for the old v1 JSON file. This now works correctly with the new v2 JSON file

Also updated the DisplayMagicianSetup to remove unneeded files to make the overall package smaller. It's now slimmed down to 12MB now, and it was 14MB. So that's a good bit of space saving!