Disabled Process Impersonation and added Profile Name to Tray Icon
Disabled the Process Impersonation logic as it was causing issues with an mscorlib.dll CTD. Have disabled as it's not strictily necessary, and it means that the games all work!

The DisplayMagician profile name is now shown when the user hovers over the system tray icon. Implements #61 for @whitestar127.

Also added some try/catch statements in the SteamLibrary IsRunning functions just-in-case :).